Dispatch cancel event to the correct target display
When screen recording is enabled, a virtual screen will be generated,
and the cancel event will be incorrectly sent to the virtual screen.
The coordinates of the cancel event are the wrong logical coordinates
after being transformed by the transform of the windows of the virtual
screen.

Bug: 301020008
Test: atest inputflinger_tests
Change-Id: Icf1301e5174ac61eeedda326545f2b68ef7b8af0
Merged-In: Icf1301e5174ac61eeedda326545f2b68ef7b8af0
Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>